wip: list files via webdav
todo: save file
check file directory etc.

+++ doris-rds/src/DorisScieboRdsConnector/Services/Storage/S3StorageService.cs
@@ -55,7 +55,7 @@ public class S3StorageService : IStorageService
                 
                 fileList.Add(new Models.File(
                     item.Key, 
-                    item.Size, 
+                    (long)item.Size, 
                     item.LastModifiedDateTime,
                     stat.ContentType,
                     item.ETag,
